--- server/FastFuel/FastFuel/Features/Common/Authorization/PermissionType.cs
@@ -1,9 +1,0 @@
-namespace FastFuel.Features.Common.Authorization;
-
-public enum PermissionType
-{
-    Create,
-    Read,
-    Update,
-    Delete
-}
