+++ server/FastFuel/FastFuel/Features/Common/Permissions/Crud/CrudPermissionType.cs
@@ -1,0 +1,9 @@
+namespace FastFuel.Features.Common.Permissions.Crud;
+
+public enum CrudPermissionType
+{
+    Create,
+    Read,
+    Update,
+    Delete
+}
