+++ server/FastFuel/FastFuel/Features/Common/Permissions/PermissionCheckAttribute.cs
@@ -1,6 +1,6 @@
 using Microsoft.AspNetCore.Mvc;
 
-namespace FastFuel.Features.Common.Authorization;
+namespace FastFuel.Features.Common.Permissions;
 
 [AttributeUsage(AttributeTargets.Method)]
 public abstract class PermissionCheckAttribute(Type type) : TypeFilterAttribute(type);
