+++ server/FastFuel/FastFuel/Features/Common/Permissions/SkipPermissionCheckAttribute.cs
@@ -1,4 +1,4 @@
-namespace FastFuel.Features.Common.Authorization;
+namespace FastFuel.Features.Common.Permissions;
 
 [AttributeUsage(AttributeTargets.Method | AttributeTargets.Class)]
 public class SkipPermissionCheckAttribute : Attribute;
