+++ server/FastFuel/FastFuel/Features/Permissions/Controllers/PermissionController.cs
@@ -1,4 +1,4 @@
-using FastFuel.Features.Common.Authorization;
+using FastFuel.Features.Common.Permissions.Crud;
 using FastFuel.Features.Permissions.Services;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http.HttpResults;
@@ -12,7 +12,7 @@ namespace FastFuel.Features.Permissions.Controllers;
 public class PermissionController(IPermissionService permissionService) : ControllerBase
 {
     [HttpGet]
-    [CrudPermissionCheck(PermissionType.Read)]
+    [CrudPermissionCheck(CrudPermissionType.Read)]
     public async Task<Results<
             Ok<List<string>>,
             UnauthorizedHttpResult,
